prescription working tests

+++ medcon/medcon.sol
@@ -24,7 +24,7 @@ contract medcon is mortal {
 		string name;
 		string username;
 		//should be a hash
-		uint password;
+		uint64 password;
 		UserRole role;
 		bool initialized;
 	}
@@ -32,8 +32,8 @@ contract medcon is mortal {
 	struct MedOrder {
 		string med_name;
 		string diagnosis;
-		uint dose_per_day;
-		uint no_of_days;
+		uint64 dose_per_day;
+		uint64 no_of_days;
 		string doctor;
 		bool initialized;
 		// if pharma has dispensed this order, mark as fulfilled
@@ -42,15 +42,15 @@ contract medcon is mortal {
 	}
 
 	struct Patient {
-		mapping(uint=>Prescription) prescriptions;
-		mapping(string=>uint) doctor_index_map;
-		uint num_of_prescriptions;
+		mapping(uint64=>Prescription) prescriptions;
+		mapping(string=>uint64) doctor_index_map;
+		uint64 num_of_prescriptions;
 		bool initialized;
 	}
 	
 	struct Prescription {
-		mapping(uint=>MedOrder) med_orders;
-		uint num_of_med_orders;
+		mapping(uint64=>MedOrder) med_orders;
+		uint64 num_of_med_orders;
 		string doctor;
 		bool initialized;
 	}
@@ -76,7 +76,7 @@ contract medcon is mortal {
 		return users[username].initialized;
 	}
 
-	function getUser(string username, string name, uint password, UserRole role) internal pure returns (User usr) {
+	function getUser(string username, string name, uint64 password, UserRole role) internal pure returns (User usr) {
 		User memory u;
 		u.name = name;
 		u.username = username;
@@ -108,7 +108,7 @@ contract medcon is mortal {
 		return p;
 	}
 
-	function getMedOrder(string med_name, uint dose_per_day, uint no_of_days, string doctor, string diagnosis) internal pure returns (MedOrder medo){
+	function getMedOrder(string med_name, uint64 dose_per_day, uint64 no_of_days, string doctor, string diagnosis) internal pure returns (MedOrder medo){
 		MedOrder memory mo;
 		mo.med_name = med_name;
 		mo.dose_per_day = dose_per_day;
@@ -129,7 +129,7 @@ contract medcon is mortal {
 		return p;
 	}
 
-	function addMedOrderToPatient(string patient, string med_name, uint dose_per_day, uint no_of_days, string doctor, string diagnosis) public {
+	function addMedOrderToPatient(string patient, string med_name, uint64 dose_per_day, uint64 no_of_days, string doctor, string diagnosis) public {
 		MedOrder memory md = getMedOrder(med_name, dose_per_day, no_of_days, doctor, diagnosis);
 		Patient storage p = patients[patient];
 		if (p.doctor_index_map[doctor] == 0) {
@@ -146,7 +146,7 @@ contract medcon is mortal {
 
 	}
 
-	function addDoctor(string username, string name, uint password, string medical_id) public {
+	function addDoctor(string username, string name, uint64 password, string medical_id) public {
 		if (users[username].initialized == true) {
 			//user already exists
 			revert();
@@ -156,7 +156,7 @@ contract medcon is mortal {
 		doctors[username] = getDoctor(medical_id);
 	}
 	
-	function addPatient(string username, string name, uint password) public {
+	function addPatient(string username, string name, uint64 password) public {
 		if (users[username].initialized == true) {
 			//user already exists
 			revert();
@@ -165,7 +165,7 @@ contract medcon is mortal {
 		patients[username] = getPatient();
 	}
 	
-	function addPharma(string username, string name, uint password, string pharma_id) public {
+	function addPharma(string username, string name, uint64 password, string pharma_id) public {
 		if (users[username].initialized == true) {
 			//user already exists
 			revert();
@@ -175,13 +175,13 @@ contract medcon is mortal {
 	}
 ///////////////////////////////////////////////////////////////////////////////////////////
 
-	function showNumOfMedOrdersByDoc(string patient, string doctor) public view returns (uint) {
+	function showNumOfMedOrdersByDoc(string patient, string doctor) public view returns (uint64) {
 		Patient storage p = patients[patient];
 		Prescription storage pr = p.prescriptions[p.doctor_index_map[doctor]];
 		return pr.num_of_med_orders;
 	}
 	
-	function showNumOfMedOrdersByIndex(string patient, uint indx) public view returns (uint) {
+	function showNumOfMedOrdersByIndex(string patient, uint64 indx) public view returns (uint64) {
 		Patient storage p = patients[patient];
 		if (indx > p.num_of_prescriptions) {
 			revert();
@@ -190,12 +190,12 @@ contract medcon is mortal {
 		return pr.num_of_med_orders;
 	}
 	
-	function showNumOfPrescriptions(string patient) public view returns (uint) {
+	function showNumOfPrescriptions(string patient) public view returns (uint64) {
 		Patient storage p = patients[patient];
 		return p.num_of_prescriptions;
 	}
 
-	function showMedOrderByDoctor(string patient, string doctor, uint index) public view returns (string, string, uint, uint, string, bool, string) {
+	function showMedOrderByDoctor(string patient, string doctor, uint64 index) public view returns (string, string, uint64, uint64, string, bool, string) {
 		Patient storage p = patients[patient];
 		Prescription storage pr = p.prescriptions[p.doctor_index_map[doctor]];
 		if (index >= pr.num_of_med_orders){
@@ -206,7 +206,7 @@ contract medcon is mortal {
 		return (m.med_name, m.diagnosis, m.dose_per_day, m.no_of_days, m.doctor, m.fulfilled, m.pharma); 
 	}
 	
-	function showMedOrderByIndex(string patient, uint pres, uint index) public view returns (string, string, uint, uint, string, bool, string) {
+	function showMedOrderByIndex(string patient, uint64 pres, uint64 index) public view returns (string, string, uint64, uint64, string, bool, string) {
 		Patient storage p = patients[patient];
 		if (!p.initialized) {
 			revert();
